+++ client/src/components/AddDrinkForm.jsx
@@ -1,86 +1,39 @@
 import { useContext, useState } from 'react';
 import axios from 'axios';
+import React from 'react';
 
 const AddDrinkForm = () => {
-  const [cocktailName, updateCocktailName] = useState('N/A');
-  const [liquors, updateLiquors] = useState([]);
-  const [ingredients, updateIngredients] = useState([]);
-  const [garnishes, updateGarnishes] = useState([]);
-  const [directions, updateDirections] = useState([]);
+  const [cocktailName, updateCocktailName] = useState('');
+  const [liquors, updateLiquors] = useState('whiskey');
+  const [ingredients, updateIngredients] = useState('');
+  const [garnishes, updateGarnishes] = useState('');
+  const [directions, updateDirections] = useState('');
 
   const postToDB = () => {
-    let garnishString ='Garnish:';
-    //do some logic to convert our garnishes array to a string. 
-    if(garnishes.length === 0) garnishString += ' none';
-    else{
-      //iterate over every garnish in our garnish array.
-      garnishes.forEach(garnish => {
-      //add each one to the garnish string
-        garnishString += ` ${garnish},`
-      });
-      //remove the final comma from the string.    
-      garnishString = garnishString.slice(0,-1);
-    }
+    //name will stay the same
+    //ingredients needs to be parsed [['ingred1','amount1'],['ingred2', 'amount2']]
+    //liquors parsed into single elemnet array --> ['Gin']
+    //garnishes string --> "Garnish: garnish1, garnish2, garnish3"
+    //directions each line parsed into arry element --> ['Step 1: ...', 'Step 2...']
+
+    const parsedName = cocktailName;
+    const parsedIngredients = ingredients.split('\n').map(e=>e.split(','));
+    const parsedGarnish = 'Garnish: ' + garnishes;
+    const parsedDirections = directions.split(`\n`);
+    const parsedLiquors = [liquors];
 
     const reqBody = {
-      name: cocktailName,
-      ingredients: ingredients,
-      garnish: garnishString,
-      directions: directions,
-      liquor: liquors
+      name: parsedName,
+      ingredients: parsedIngredients,
+      garnish: parsedGarnish,
+      directions: parsedDirections,
+      liquor: parsedLiquors
     }
+    console.log('req body:', reqBody);
     axios.post('http://localhost:8080/', reqBody)
     .then(res => location.reload())
     .catch(err => console.log(err));
   }
-
-  const addLiquor = (liquor) =>{
-    updateLiquors([...liquors, liquor])
-  }
-
-  const liquorsElements = [];
-  liquors.forEach((liquor, index) => {
-    liquorsElements.push(
-      <li key={index} className='liquor'>{liquor}</li>
-    )
-  })
-
-
-  const addIngredients = (amount, ingredient) => {
-    updateIngredients([...ingredients, [amount+' ',ingredient]]);
-  }
-  const ingredientElments = [];
-  ingredients.forEach((ingredientArray, index) => {
-    ingredientElments.push(
-      <li key = {index} className = 'ingredient'>{ingredientArray[0]}, {ingredientArray[1]}</li>
-    )
-  })
-
-
-  const addGarnish = (garnish) => {
-    updateGarnishes([...garnishes, garnish]);
-  }
-
-  const garnishElements = [];
-  garnishes.forEach((garnish, index) => {
-    garnishElements.push(
-      <li key = {index} className = 'garnish'>{garnish}</li>
-    )
-  })
-
-  const addDirection = (direction) => {
-    const step = directions.length + 1;
-    direction = `STEP ${step}: ` + direction;
-    updateDirections([...directions, direction]);
-  }
-
-  const directionsElements = [];
-  directions.forEach((direction, index) => {
-    directionsElements.push(
-      <li key = {index} className = 'direction'>{direction}</li>
-    )
-  })
-
 
   const categories = [
    
@@ -102,8 +55,7 @@ const AddDrinkForm = () => {
       {/* name is a string */}
       <div id= 'addIngredientsDiv'>
       <label htmlFor='name' style={{display: 'block'}}>Name of your cocktail:</label>
-      <input type='text' name='name' id='name'></input>
-      <button style={{display: 'block'}} onClick = {() => updateCocktailName(document.getElementById('name').value)}>Submit Name</button>
+      <input type='text' name='name' id='name' onChange = {() => updateCocktailName(document.getElementById('name').value)}></input>
      
       <br/>
       
@@ -113,7 +65,7 @@ const AddDrinkForm = () => {
       <select
           id="liquor"
           // value={selected}
-          // onChange={handleSelectChange}
+          onChange={() => updateLiquors([document.getElementById('liquor').value])}
         >
           {categories.map((cat) => (
             <option key={cat} value={cat}>
@@ -121,41 +73,45 @@ const AddDrinkForm = () => {
             </option>
           ))}
         </select>
-      <button style={{display: 'block'}} onClick = {() => addLiquor(document.getElementById('liquor').value)}>Choose your Liquor(s)</button>
       <br/>
 
       {/* ingredients is an array of array of strings. example: */}
       {/*ingredients: [ ['lime cordial', '50ml'], ['gin', '50ml'] ]*/}
-      <label style={{display: 'block'}} htmlFor='ingredients'>Ingredient Amount:</label>
-      <input type='text' name='ingredient Amount' id='ingredient-amount'></input>
-      <label style={{display: 'block'}} htmlFor='ingredients'> Ingredient Name:</label>
-      <input type='text' name='ingredient Items' id='ingredient-items'></input>
-      <button 
+      <label style={{display: 'block'}} htmlFor='ingredients'>Ingredients (ingredient, amount)</label>
+      <textarea placeholder={'INGREDIENT1, AMOUNT1\nINGREDIENT2, AMOUNT2'} type='text' name='ingredients' id='ingredients' onChange = {() => {updateIngredients(document.getElementById('ingredients').value)}}></textarea>
+      {/* <button 
         style={{display: 'block'}} 
         onClick ={() => addIngredients(document.getElementById('ingredient-amount').value, document.getElementById('ingredient-items').value)}
         >
         Submit ingredient
       </button>
-      
+       */}
       <br/>
 
       {/*Garnish is an array of strings i.e.: garnish: ['lime', 'edible flower'] */}
       <label style={{display: 'block'}} htmlFor='garnish'>Garnish :</label>
-      <input type='text' name='garnish' id='garnishInput'></input>
-      <button style={{display: 'block'}} onClick ={() => {addGarnish(document.getElementById('garnishInput').value)}}>Add the garnish</button>
+      <input type='text' name='garnish' placeholder={'garnish1, garnish2, etc.'} id='garnishInput' onChange = {() => {updateGarnishes(document.getElementById('garnishInput').value)}}></input>
+      {/* <button style={{display: 'block'}} onClick ={() => {addGarnish(document.getElementById('garnishInput').value)}}>Add the garnish</button> */}
       <br/>
     
       {/*Directions is a string with special formatting i.e.
       'STEP 1 .... STEP 2......
       */}
       <label style={{display: 'block'}} htmlFor='liquor'>Directions :</label>
-      <input type='text' name='directions' id='directions'></input>
-      <button style={{display: 'block'}} onClick = {() => addDirection(document.getElementById('directions').value)}>Add the directions</button>
-    
+      <textarea placeholder = {`Step 1: first step \nStep 2: second step \nStep 3: ...`}type='text' name='directions' id='directions' onChange = {() => {updateDirections(document.getElementById('directions').value)}}></textarea>
+      {/* <button style={{display: 'block'}} onClick = {() => addDirection(document.getElementById('directions').value)}>Add the directions</button> */}
+<br/>
+      <button onClick = {() => {
+
+        postToDB()
+        location.reload()
+        ;
+      }
+      }>Submit your Cocktail!</button>
     
     </div>
 
-    <div id='newDrinkState' className='Modal'>
+    {/* <div id='newDrinkState' className='Modal'>
       <p>Cocktail name: </p>
       <h2>{cocktailName}</h2>
       <p>Liquors:</p>
@@ -167,7 +123,7 @@ const AddDrinkForm = () => {
       <p>Directions:</p>
       {directionsElements}
       <button onClick = {() => postToDB()}>Click HERE to submit custom cocktail</button>
-    </div>
+    </div> */}
   </div>
   )
 };
