+++ client/src/components/MobileFilters.jsx
@@ -38,7 +38,7 @@ function MobileFilters({ selected, select }) {
             <h2>Don't like what you see?</h2>
               <ul className="filters">
                 <li>
-                  <button name='filter' className="filter-btn" id="mobile-button">Add Recipie</button>
+                  <button name='filter' className="filter-btn" id="mobile-button">Add Recipe</button>
                 </li>
               </ul>
             </div>
