+++ client/src/index.css
@@ -180,6 +180,7 @@ button:focus-visible {
 .card-box img {
   transition: all 0.8s ease-in-out;
   transform: scale(1);
+  background-color: #f9f5cf;
 }
 .card-box:hover img {
   transform: scale(1.1);
@@ -463,6 +464,7 @@ header h2 {
     flex-direction: row;
     height: auto;
     width: auto;
+    background-color: ;
   }
   #modal .card-box img {
     display: none;
