+++ server/controllers/spiritController.js
@@ -8,14 +8,14 @@ const baseError = {
 
 const spiritController = {};
 
-spiritController.createSpirit = asyncHandler((req, res) => {
+spiritController.createSpirit = asyncHandler( async (req, res) => {
   const { name, ingredients, liquor, directions, garnish } = req.body;
-  console.log('these are the req body things:')
-  console.log('name:', name);
-  console.log('ingredients:', ingredients);
-  console.log('liquor:', liquor);
-  console.log('directions:', directions);
-  console.log('garnish:', garnish);
+  // console.log('these are the req body things:')
+  // console.log('name:', name);
+  // console.log('ingredients:', ingredients);
+  // console.log('liquor:', liquor);
+  // console.log('directions:', directions);
+  // console.log('garnish:', garnish);
   const spirit = await Spirit.create({
     name,
     ingredients,
@@ -39,9 +39,8 @@ spiritController.createSpirit = asyncHandler((req, res) => {
 });
 
 spiritController.deleteSpirit = asyncHandler(async (req,res,next) => {
-  console.log('req.body', req.body)
+  // console.log('req.body', req.body)
   const { name } = req.body;
-  console.log('about to delete:', name)
   await Spirit.findOneAndDelete({ name })
   return next();
 });
@@ -53,16 +52,6 @@ spiritController.getSpirits = asyncHandler(async (req, res, next) => {
 
   //Mongo and/or Express did not like me using const here, why?
   res.locals.spirits = await Spirit.find();
-  return next();
-});
-
-spiritController.deleteSpirit = asyncHandler(async (req, res, next) => {
-  await Spirit.findOneAndDelete({ _id: res.locals.id });
-  return next();
-});
-
-spiritController.updateSpirit = asyncHandler(async (req, res, next) => {
-  await Spirit.findOneAndUpdate({ _id: res.locals.id });
   return next();
 });
